services/cloud-adapter-service/requirements.txt
paho-mqtt
requests
Flask
PyYAML 